event_management_service/pyproject.toml
[tool.poetry]
name = "event-service"
version = "0.1.0"
description = "A microservice for managing events in Lisbon, including event creation, retrieval, updating, and deletion. It serves as the core service for storing and providing event data to other services and users"
authors = ["Yulia Ioffe <https://github.com/leanor13>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
setuptools = "^75.2.0"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
flake8 = "^7.1.1"
black = "^24.8.0"
pre-commit = "^4.0.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
